:tractor: code refactoring on plugins.py
