+++ moban/plugins/__init__.py
@@ -7,7 +7,8 @@ import moban.reporter as reporter
 from moban import utils, constants, exceptions
 from moban.plugins.strategy import Strategy
 from moban.hashstore import HASH_STORE
-from moban.data_loaders.manager import AnyDataLoader
+from moban.plugins.context import Context
+from moban.plugins.library import LIBRARIES
 
 BUILTIN_EXENSIONS = [
     "moban.jinja2.engine",
@@ -16,21 +17,12 @@ BUILTIN_EXENSIONS = [
 ]
 
 
-class LibraryManager(PluginManager):
-    def __init__(self):
-        super(LibraryManager, self).__init__(constants.LIBRARY_EXTENSION)
-
-    def resource_path_of(self, library_name):
-        library = self.get_a_plugin(library_name)
-        return library.resources_path
-
-
 class BaseEngine(object):
     def __init__(
         self, template_dirs, context_dirs, engine_cls, engine_extensions=None
     ):
         template_dirs = list(expand_template_directories(template_dirs))
-        verify_the_existence_of_directories(template_dirs)
+        utils.verify_the_existence_of_directories(template_dirs)
         context_dirs = expand_template_directory(context_dirs)
         self.context = Context(context_dirs)
         self.template_dirs = template_dirs
@@ -167,26 +159,7 @@ class EngineFactory(PluginManager):
         raise exceptions.NoThirdPartyEngine(key)
 
 
-LIBRARIES = LibraryManager()
 ENGINES = EngineFactory()
-LOADERS = AnyDataLoader()
-
-
-def load_data(base_dir, file_name):
-    abs_file_path = utils.search_file(base_dir, file_name)
-    data = LOADERS.get_data(abs_file_path)
-    if data is not None:
-        parent_data = None
-        if base_dir and constants.LABEL_OVERRIDES in data:
-            parent_data = load_data(
-                base_dir, data.pop(constants.LABEL_OVERRIDES)
-            )
-        if parent_data:
-            return utils.merge(data, parent_data)
-        else:
-            return data
-    else:
-        return None
 
 
 def expand_template_directories(dirs):
@@ -227,47 +200,5 @@ def expand_template_directory(directory):
     return translated_directory
 
 
-class Context(object):
-    def __init__(self, context_dirs):
-        verify_the_existence_of_directories(context_dirs)
-        self.context_dirs = context_dirs
-        self.__cached_environ_variables = dict(
-            (key, os.environ[key]) for key in os.environ
-        )
-
-    def get_data(self, file_name):
-        try:
-            data = load_data(self.context_dirs, file_name)
-            utils.merge(data, self.__cached_environ_variables)
-            return data
-        except (IOError, exceptions.IncorrectDataInput) as exception:
-            # If data file doesn't exist:
-            # 1. Alert the user of their (potential) mistake
-            # 2. Attempt to use environment vars as data
-            reporter.report_warning_message(str(exception))
-            reporter.report_using_env_vars()
-            return self.__cached_environ_variables
-
-
 def make_sure_all_pkg_are_loaded():
     scan_plugins_regex(constants.MOBAN_ALL, "moban", None, BUILTIN_EXENSIONS)
-
-
-def verify_the_existence_of_directories(dirs):
-    if not isinstance(dirs, list):
-        dirs = [dirs]
-
-    for directory in dirs:
-        if os.path.exists(directory):
-            continue
-        should_I_ignore = (
-            constants.DEFAULT_CONFIGURATION_DIRNAME in directory
-            or constants.DEFAULT_TEMPLATE_DIRNAME in directory
-        )
-        if should_I_ignore:
-            # ignore
-            pass
-        else:
-            raise exceptions.DirectoryNotFound(
-                constants.MESSAGE_DIR_NOT_EXIST % os.path.abspath(directory)
-            )
