+++ tests/integration_tests/test_command_line_options.py
@@ -14,7 +14,7 @@ class TestCustomOptions:
         with open(self.config_file, "w") as f:
             f.write("hello: world")
         self.patcher1 = patch(
-            "moban.plugins.verify_the_existence_of_directories"
+            "moban.utils.verify_the_existence_of_directories"
         )
         self.patcher1.start()
 
@@ -69,7 +69,7 @@ class TestOptions:
         with open(self.config_file, "w") as f:
             f.write("hello: world")
         self.patcher1 = patch(
-            "moban.plugins.verify_the_existence_of_directories"
+            "moban.utils.verify_the_existence_of_directories"
         )
         self.patcher1.start()
 
@@ -129,7 +129,7 @@ class TestNoOptions:
         with open(self.data_file, "w") as f:
             f.write("hello: world")
         self.patcher1 = patch(
-            "moban.plugins.verify_the_existence_of_directories"
+            "moban.utils.verify_the_existence_of_directories"
         )
         self.patcher1.start()
 
@@ -218,7 +218,7 @@ class TestNoOptions2:
         with open(self.data_file, "w") as f:
             f.write("hello: world")
         self.patcher1 = patch(
-            "moban.plugins.verify_the_existence_of_directories"
+            "moban.utils.verify_the_existence_of_directories"
         )
         self.patcher1.start()
 
@@ -256,7 +256,7 @@ class TestCustomMobanFile:
         with open(self.data_file, "w") as f:
             f.write("hello: world")
         self.patcher1 = patch(
-            "moban.plugins.verify_the_existence_of_directories"
+            "moban.utils.verify_the_existence_of_directories"
         )
         self.patcher1.start()
 
@@ -291,7 +291,7 @@ class TestTemplateOption:
             os.path.join("tests", "fixtures", ".moban.yml"), self.config_file
         )
         self.patcher1 = patch(
-            "moban.plugins.verify_the_existence_of_directories"
+            "moban.utils.verify_the_existence_of_directories"
         )
         self.patcher1.start()
 
@@ -322,7 +322,7 @@ class TestTemplateOption:
         os.unlink(self.config_file)
 
 
-@patch("moban.plugins.verify_the_existence_of_directories")
+@patch("moban.utils.verify_the_existence_of_directories")
 def test_duplicated_targets_in_moban_file(fake_verify):
     config_file = "duplicated.moban.yml"
     copyfile(os.path.join("tests", "fixtures", config_file), ".moban.yml")
@@ -385,7 +385,7 @@ class TestComplexOptions:
         with open(self.data_file, "w") as f:
             f.write("hello: world")
         self.patcher1 = patch(
-            "moban.plugins.verify_the_existence_of_directories"
+            "moban.utils.verify_the_existence_of_directories"
         )
         self.patcher1.start()
